Do not send the whole request.headers recieved by Kibana to the ES
The requestHeaders param of toolkit.authenticated() defines the headers
that are added to requests sent to ES.

It is wrong to send ALL headers received by Kibana (which includes
things like host, cookie, content-length, connection keep-alive) again
to ES, as these are completely unrelated requests. At maximum, the
authorization header may be passed on unmodified.

Especially the host header causes TLS error because the SAN of the
backend certificate usually does not match the name of the host Kibana
is running on.